+++ tests/api/test-dev-lightfunc.c
@@ -804,6 +804,8 @@ key: isPrototypeOf
 key: propertyIsEnumerable
 key: __defineGetter__
 key: __defineSetter__
+key: __lookupGetter__
+key: __lookupSetter__
 top: 1
 enum own
 top: 1